content/html/document/test/test_bug871161.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=871161
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 871161</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 871161 **/
  SimpleTest.waitForExplicitFinish();

  window.onmessage = function(e) {
    is(e.data, "windows-1252", "Wrong charset");
    e.source.close();
    SimpleTest.finish();
  }

  function run() {
    window.open("file_bug871161-1.html");
  }

  </script>
</head>
<body onload="run();">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=871161">Mozilla Bug 871161</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
